library/express-pool-kafka/pyproject.toml
[tool.poetry]
name = "express-pool-kafka"
version = "1.0.4"
description = ""
authors = ["yjsong <yjsong@interezen.co.kr>"]
readme = "README.md"
packages = [{ include = "express_pool_kafka" }]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^6.0.1"
kafka-python = "^2.0.2"
express-utils = {path = "../../library/express-utils", develop = true}

[[tool.poetry.source]]
name = "pypi-group"
url = "http://192.168.124.238:8001/repository/pypi-group/simple/"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

